+++ lib/fossa/views/controller.py
@@ -7,6 +7,11 @@ def node_summary(governor):
     @param governor: instance of :class:`fossa.control.governor.Governor`
     @return: dict with keys-
         - "recent_completed_tasks" - list of dict with details of task
+                additional 'status' key which can have values-
+                - running
+                - failed
+                - complete
+                - unknown (indicates something missing in the code)
         - "running_tasks"
         - "node_info" - dict
     """
@@ -22,15 +27,28 @@ def node_summary(governor):
         summary = asdict(t["task_spec"])
         summary["started"] = t["started"]
         summary["finished"] = t["finished"]
-        summary["results"] = t["result_spec"].task_message
+        summary["results"] = json.loads(t["result_spec"].task_message)
+
+        if summary["results"]["type"] == "TaskComplete":
+            summary["status"] = "complete"
+        elif summary["results"]["type"] == "TaskFailed":
+            summary["status"] = "failed"
+        else:
+            summary["status"] = "unknown"
+
         previous_tasks.append(summary)
+
+    previous_tasks.sort(key=lambda t: t["finished"], reverse=True)
 
     current_tasks = []
     for task_id, process_details in governor.process_table.items():
         process_extract = asdict(process_details["task_spec"])
         process_extract["task_id"] = task_id
         process_extract["started"] = process_details["started"]
+        process_extract["status"] = "running"
         current_tasks.append(process_extract)
+
+    current_tasks.sort(key=lambda t: t["started"], reverse=False)
 
     ns = {
         "running_tasks": current_tasks,
@@ -47,27 +65,16 @@ def task_summary(governor, task_id):
 
     @return (dict) or None if task_id not known
         keys from :class:`TaskMessage`
-        with additional 'status' key which can have values-
-        - running
-        - failed
-        - complete
     """
 
     ns = node_summary(governor)
 
     for t in ns["running_tasks"]:
         if t["task_id"] == task_id:
-            t["status"] = "running"
             return t
 
     for t in ns["recent_completed_tasks"]:
         if t["task_id"] == task_id:
-            results = json.loads(t["results"])
-
-            if results["type"] == "TaskComplete":
-                t["status"] = "complete"
-            else:
-                t["status"] = "failed"
             return t
 
-    return None  # not known
+    return None  # task not known